requirements.txt
opencv-python
Pillow
numpy
torch
torchvision
tqdm
flask
flask-socketio
mediapipe
